WIP - refactoring oauth functions to enable refresh functionality
