+++ backend/open_webui/apps/webui/main.py
@@ -44,7 +44,7 @@ from open_webui.config import (
     WEBHOOK_URL,
     WEBUI_AUTH,
     WEBUI_BANNERS,
-    AppConfig, OAUTH_ALLOWED_ROLES, OAUTH_ADMIN_ROLES,
+    AppConfig,
 )
 from open_webui.env import (
     WEBUI_AUTH_TRUSTED_EMAIL_HEADER,
